Move `GuildWelcomeScreen` and dependant types into a separate module
It slices off a bit of the enormous `guild/mod.rs` file.
